Move RemoteChannelViewSet into its own module
Distinct from viewsets/remote.py, which holds the RemoteMixin proxying pair.
